test(searchbar): migrate
