+++ core/src/components/searchbar/test/basic/searchbar.e2e.ts-snapshots/searchbar-color-ios-ltr-Mobile-Chrome-linux.png
