+++ core/src/components/searchbar/test/basic/searchbar.e2e.ts-snapshots/searchbar-color-md-ltr-Mobile-Firefox-linux.png
