+++ core/src/components/searchbar/test/basic/searchbar.e2e.ts-snapshots/searchbar-ios-ltr-Mobile-Safari-linux.png
