+++ core/src/components/searchbar/test/basic/searchbar.e2e.ts-snapshots/searchbar-ios-rtl-Mobile-Safari-linux.png
