+++ core/src/components/searchbar/test/basic/searchbar.e2e.ts-snapshots/searchbar-md-ltr-Mobile-Chrome-linux.png
