+++ core/src/components/searchbar/test/basic/searchbar.e2e.ts-snapshots/searchbar-md-rtl-Mobile-Safari-linux.png
