+++ core/src/components/searchbar/test/basic/searchbar.e2e.ts-snapshots/searchbar-search-icon-ios-ltr-Mobile-Safari-linux.png
